% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{create_labels_dictionary}
\alias{create_labels_dictionary}
\title{Crea un diccionario de  valores de una columna.}
\usage{
create_labels_dictionary(df, col_codes, col_labels)
}
\arguments{
\item{df}{Un objeto de clase "data.frame" con los datos de una tabla en SIMEL.}

\item{col_codes}{Valor de clase "character"  con la columna con los códigos.}

\item{col_labels}{Valor de clase "character" con la columna con las labels.}
}
\value{
Un vector de clase "character" tipo "named vector" (diccionario).
}
\description{
Crea un diccionario de  valores de una columna.
}
\examples{
tabla <- data.frame(levels = 1:3, labels = c("uno", "dos", "tres"))
create_labels_dictionary(tabla, "levels", "labels")
}
