% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{desagregaciones}
\alias{desagregaciones}
\title{Desagregaciones de SIMEL}
\usage{
desagregaciones()
}
\value{
Un vector de clase "character" con los nombres de las columnas de
desagregación.
}
\description{
Desagregaciones de SIMEL
}
