% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_data}
\alias{get_data}
\title{Obtener data SIMEL}
\usage{
get_data(dataflow, ...)
}
\arguments{
\item{dataflow}{Un único valor de clase "character" con el nombre del
dataflow (id en tabla indicadores).}

\item{...}{Otros argumentos de readSMDX}
}
\value{
Un objeto de clase "data.frame" con los datos de una tabla en SIMEL.
}
\description{
Descarga información de una tabla de datos publicada en SIMEL.
}
\examples{
get_data("TP_SEXO")
}
