% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_data_url}
\alias{get_data_url}
\title{Obtener url API SIMEL Chile (data)}
\usage{
get_data_url(dataflow)
}
\arguments{
\item{dataflow}{Un único valor de clase "character" con el nombre del dataflow (id en tabla indicadores).}
}
\value{
Un único valor de clase "character" con la url de la API para la data.
}
\description{
Obtener url API SIMEL Chile (data)
}
\examples{
get_data_url("TP_SEXO")
}
