% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{get_dsd_url}
\alias{get_dsd_url}
\title{Obtener url API SIMEL Chile (metadata)}
\usage{
get_dsd_url(dataflow)
}
\arguments{
\item{dataflow}{Un único valor de clase "character" con el nombre del
dataflow (id en tabla indicadores).}
}
\value{
Un único valor de clase "character" con la url de la API para la
metadata.
}
\description{
Obtener url API SIMEL Chile (metadata)
}
\examples{
get_dsd_url("TP_SEXO")
}
