% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{identificar_trimestre_movil}
\alias{identificar_trimestre_movil}
\title{Identifica trimestre movil}
\usage{
identificar_trimestre_movil(df)
}
\arguments{
\item{df}{Objeto de clase "data.frame" proveniente de SIMEL.}
}
\value{
Objeto clase "Logical" con valor TRUE cuando el data.frame viene
con fechas en formato trimestre movil.
}
\description{
Identifica si la columna TIME_PERIOD  viene en formato de trimestre movil.
}
\examples{
tbl <- data.frame(TIME_PERIOD = "2016-05-01/P3M")
identificar_trimestre_movil(tbl)
}
