% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{labelizar_df}
\alias{labelizar_df}
\title{Labeliza una tabla de SIMEL}
\usage{
labelizar_df(df)
}
\arguments{
\item{df}{Un objeto de clase "data.frame" con los datos de una tabla en SIMEL.}
}
\value{
Un objeto de clase "data.frame" con las variables categóricas
en formato haven::labelled.
}
\description{
Transforma en variables haven::labelled() las columnas que tienen una
label asociada en las tablas SIMEL (excepto por columna DECIMAL).
}
\examples{
tabla <- data.frame(numeros = 1:3, numeros_label.es = c("uno", "dos", "tres"))
labelizar_df(tabla)
}
