% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{lista_indicadores}
\alias{lista_indicadores}
\title{Obtener lista indicadores.}
\usage{
lista_indicadores()
}
\value{
Objeto clase "data.frame" con la lista indicadores de SIMEL.
}
\description{
Retorna la lista de indicadores.
}
\examples{
indicadores <- lista_indicadores()
}
