% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{simplificar_tabla}
\alias{simplificar_tabla}
\title{Simplificar tablas SUSESO}
\usage{
simplificar_tabla(df, redondear = TRUE)
}
\arguments{
\item{df}{Objeto de clase "data.frame" proveniente de SIMEL.}

\item{redondear}{Valor lógico.}
}
\value{
Un objeto de clase "data.frame" simplificado.
}
\description{
Simplifica las tablas de SUSESO seleccionando las  columnas más importantes. Además, redondea  los valores al número de decimales establecidos en cada tablas.
}
\examples{
\dontrun{
df <- get_data("TP_SEXO")
df_v2 <- simplificar_tabla(df)
}
}
