% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tabla_indicadores}
\alias{tabla_indicadores}
\title{Tabla con los indicadores de SIMEL con sus respectivos IDs.}
\format{
Un data frame con 259 filas y 4 columnas:
\describe{
\item{id_number }{Número de la tabla.}
\item{id}{Nombre en código de la tabla (dataflow).}
\item{hp}{Fuente de los datos.}
\item{Tabla}{Nombre de la tabla de datos.}

}
}
\usage{
tabla_indicadores
}
\description{
Tabla con los indicadores de SIMEL con sus respectivos IDs.
}
\keyword{datasets}
