% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{timeperiod_to_date}
\alias{timeperiod_to_date}
\title{Transformar timeperiod a clase "date"}
\usage{
timeperiod_to_date(df)
}
\arguments{
\item{df}{Objeto de clase "data.frame" proveniente de SIMEL.}
}
\value{
Un objeto de clase "data.frame" con la columna TIME_PERIOD como
clase "date".
}
\description{
Transforma la variable TIME_PERIOD a clase date.
}
\examples{
\dontrun{
df <- get_data("TP_SEXO")
df_v2 <- timeperiod_to_date(df)
}
}
