% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{to_labelled}
\alias{to_labelled}
\title{Labeliza una columna}
\usage{
to_labelled(df, col_name, labels_dictionary)
}
\arguments{
\item{df}{Un objeto de clase "data.frame" con los datos de una tabla en SIMEL.}

\item{col_name}{Objeto "character" con el nombre de la columna a transformar.}

\item{labels_dictionary}{Objeto tipo named vector  utilizado como diccionario.}
}
\value{
Objeto de clase "data.frame" con la columna  en formato
haven::labelled.
}
\description{
Permite transformar una columna a haven::labelled a partir de un diccionario.
}
\examples{
tabla = data.frame(valores = 1:3)
diccionario = c(1 = "uno", 2 = "dos", 3 = "tres")
to_labelled(tabla, "valores", diccionario)
}
