% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ver_lista_indicadores}
\alias{ver_lista_indicadores}
\title{Ver lista de indicadores}
\usage{
ver_lista_indicadores()
}
\value{
Print de la lista de indicadores.
}
\description{
Imprime la lista de indicadores.
}
\examples{
ver_lista_indicadores()
}
